resolve conflicts
