+++ pom.xml
@@ -60,7 +60,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-redis</artifactId>
-            <scope>runtime</scope>
         </dependency>
 
         <dependency>
