+++ src/main/java/com/nhnacademy/edu/springboot/minidooray/gateway/config/RedisConfig.java
@@ -10,6 +10,7 @@ import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactor
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
+
 @Configuration
 @ConfigurationProperties(prefix = "spring.redis")
 @Setter
