--- src/main/java/com/nhnacademy/edu/springboot/minidooray/gateway/config/RestTemplateConfig.java
@@ -1,25 +1,0 @@
-package com.nhnacademy.edu.springboot.minidooray.gateway.config;
-
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.BufferingClientHttpRequestFactory;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
-import org.springframework.http.converter.StringHttpMessageConverter;
-import org.springframework.web.client.RestTemplate;
-
-import java.nio.charset.StandardCharsets;
-import java.time.Duration;
-
-@Configuration
-public class RestTemplateConfig {
-    @Bean
-    public RestTemplate restTemplate(RestTemplateBuilder restTemplateBuilder) {
-        return restTemplateBuilder
-                .requestFactory(() -> new BufferingClientHttpRequestFactory(new SimpleClientHttpRequestFactory()))
-                .setConnectTimeout(Duration.ofMillis(5L * 1000))
-                .setReadTimeout(Duration.ofMillis(5L * 1000))
-                .additionalMessageConverters(new StringHttpMessageConverter(StandardCharsets.UTF_8))
-                .build();
-    }
-}
